requirements.txt
google-generativeai
openai
pandas
requests
arxiv
semanticscholar
pdfminer.six
PyPDF2
python-dotenv
